Move archiveStore/utils/hierarchicalLayers to archiveStore/differentialState/hierarchicalLayers
