+++ packages/beacon-node/src/chain/archiveStore/differentialState/hierarchicalLayers.ts
