+++ packages/beacon-node/src/chain/archiveStore/differentialState/operation.ts
@@ -4,8 +4,8 @@ import {Slot} from "@lodestar/types";
 import {Logger} from "@lodestar/utils";
 import {IBeaconDb} from "../../../db/interface.js";
 import {IStateDiffCodec} from "../interface.js";
-import {HierarchicalLayers} from "../utils/hierarchicalLayers.js";
 import {replayBlocks} from "../utils/replayBlocks.js";
+import {HierarchicalLayers} from "./hierarchicalLayers.js";
 import {BeaconStateSnapshot} from "./ssz.js";
 import {getStateDifferentials, replayStateDifferentials} from "./stateDifferential.js";
 import {beaconStateBytesToSnapshot, getStateSnapshot, snapshotToBeaconStateBytes} from "./stateSnapshot.js";
